chore: moving pixel to store-sdk and renaming it to analytics
