+++ packages/gatsby-plugin-pixel-facebook/src/pixel/index.tsx
@@ -1,6 +1,7 @@
 import React, { Fragment, useMemo } from 'react'
 import type { FC } from 'react'
-import { useLazyScript, usePixelEvent, once } from '@vtex/gatsby-theme-store'
+import { useAnalyticsEvent } from '@vtex/store-sdk'
+import { useLazyScript, once } from '@vtex/gatsby-theme-store'
 
 import type { FBPixelProviderProps } from './types'
 import { handler } from './handler'
@@ -59,7 +60,7 @@ const Provider: FC<FBPixelProviderProps> = ({
     })
 
     // Register Pixel event handler
-    usePixelEvent(handler)
+    useAnalyticsEvent(handler)
   }
 
   return <Fragment>{children}</Fragment>
