+++ packages/gatsby-theme-store/src/index.ts
@@ -90,10 +90,6 @@ export { setRegion } from './sdk/region/setRegion'
 
 export { useToast } from './sdk/toast/useToast'
 
-export { usePixelSendEvent } from './sdk/pixel/usePixelSendEvent'
-export { usePixelEvent } from './sdk/pixel/usePixelEvent'
-export type { PixelEvent } from './sdk/pixel/pixel'
-export type { PixelEventHandler } from './sdk/pixel/usePixelEvent'
 export * from './sdk/pixel/events'
 
 export { once } from './sdk/once'
